Version feature completed
in this feature, the user can tap on version send all the preferences and trips saved to the server which the developer can download and create a clone of user's app settings. The developer will copy those files and paste it into the document directory of the app and run the app.